fix : StringParser class 로 기능 분리 requestHandler 에 구현되어 있던 url,extension 에 대해 파싱해주는 부분을 따로 클래스에 구현
