+++ src/main/java/utils/Utility.java
@@ -2,12 +2,7 @@ package utils;
 
 public class Utility {
 
-    public String stringSplit(String line) {
-        String[] token = line.split(" ");
-        String indexString = token[1];
-        return indexString;
 
-    }
 
 
 }
